+++ jpype/__init__.py
@@ -52,7 +52,7 @@ __all__.extend(_jclass.__all__)  # type: ignore[name-defined]
 __all__.extend(_jcustomizer.__all__)  # type: ignore[name-defined]
 __all__.extend(_gui.__all__)  # type: ignore[name-defined]
 
-__version__ = "1.5.2.dev0"
+__version__ = "1.5.2"
 __version_info__ = __version__.split('.')