+++ doc/CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog
 This changelog *only* contains changes from the *first* pypi release (0.5.4.3) onwards.
 
 Latest Changes:
+- **1.5.2 - 2025-01-20**
 - **1.5.2.dev0 - 2024-11-18**
 
   - Roll back agent change due to misbehaving JVM installs.